requirements.txt
asgiref==3.4.1
Django==3.2.10
django-cors-headers==3.10.1
django-environ==0.8.1
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2-binary==2.9.2
pytz==2021.3
sqlparse==0.4.2
typing-extensions==4.0.1
